@@ -8,8 +8,8 @@ from errno import EINVAL
 from sys import argv
 from uuid import uuid4
 from time import sleep
+from json import loads
 
-from .._bytesjson import loads
 from .._output import MemoryLogger
 from .._message import TASK_UUID_FIELD
 from .. import start_action, Message, write_traceback
